# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PostKind = typeof PostKind[keyof typeof PostKind];


export const PostKind = {
  scene: 'scene',
  dungeon: 'dungeon',
} as const;

export type PostPayload = { [key: string]: unknown };

export interface Post {
  id: number;
  kind: PostKind;
  name: string;
  ownerId: string;
  authorName: string;
  payload: PostPayload;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreatePostPayload = { [key: string]: unknown };

export interface CreatePost {
  kind: PostKind;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  payload: CreatePostPayload;
  isPublic?: boolean;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  surface: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  surface: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiConversationInput {
  title: string;
  surface?: string;
}

export type OpenaiMessageInputToolsItem = { [key: string]: unknown };

export interface OpenaiMessageInput {
  /** @minLength 1 */
  content: string;
  system?: string;
  tools?: OpenaiMessageInputToolsItem[];
}

export interface ApiError {
  error: string;
}

export interface WalletInfo {
  address: string;
  chain: string;
  createdAt: string;
  updatedAt: string;
}

export interface LinkedWallet {
  wallet: WalletInfo | null;
}

export interface WalletNonceInput {
  /** The public wallet address (base58 for Solana) being linked */
  address: string;
}

export interface WalletNonce {
  nonce: string;
  /** The exact UTF-8 message the wallet must sign */
  message: string;
}

export interface WalletLinkInput {
  address: string;
  /** base58-encoded signature of the issued message */
  signature: string;
}

export type ListPostsParams = {
kind?: PostKind;
};

export type ListOpenaiConversationsParams = {
surface?: string;
};